ref(feedback): Refactor `<MessageSection>` into two parts
This should make it make more sense, easier to find things because its not intuitive that timeSince is inside the Message bit
